requirements.txt
Flask==3.0.3
Pillow==10.4.0
ultralytics==8.3.13
Werkzeug==3.0.4